refactor: make author functions composables
Allows reuse on individual news articles.
